Refactor ContactForm and ListView for better functionality
Adjusted the methods within ContactForm and ListView Java files to enhance functionality and efficiency. Renamed the 'createButtonLayout' method to 'createButtonsLayout', improving readability and better reflecting its purpose. Also moved the keycode escape shortcut to the cancel button, and added listener events for save, delete, and cancel buttons for better user interactions.

Configured the 'ListView' file to make better use of the service, including the methods for saving and deleting contacts. Organised the 'configureGrid' and 'editContact' methods for more efficient function implementations. Created a 'closeEditor' method to hide the editor view when not in use, enhancing the user interface.

Added responsive CSS to hide toolbar and contact grid in smaller screens for a better user experience in different device sizes.

These changes were made to streamline the user experience and improve system efficiency.