Update ListView route and add new features
Updated the ListView route to include MainLayout as its layout. Added @Formula in the company entity to count the company's employees, which will be used in the new "Dashboard" feature. Implemented styles for highlighted texts in styles.css. Additionally, two new files for the "Dashboard" and "MainLayout" features were added.

These changes are implemented to improve the user experience by providing a more comprehensive dashboard and better navigation.